backend/pyproject.toml
[project]
name = "llm-legal-advisory-council"
version = "0.1.0"
description = "AI-powered Legal Advisory Council Backend"
authors = [
    {name = "Legal Council Team", email = "dev@legalcouncil.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.17",

    # Database
    "sqlalchemy>=2.0.35",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",

    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",  # bcrypt 4.x required for passlib compatibility

    # HTTP Client
    "httpx>=0.28.0",

    # LLM APIs
    "openai>=1.55.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.0",

    # Storage
    "boto3>=1.35.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
